set the canonical without page and rel=prev|next in head
